Refactor api route into package and include a new route to list available bins
